Heap debug
- Incorrect comparison
- Unappropriate compiler options (for windows)
- Some incorrect types

+++ Src/node.h
@@ -10,7 +10,7 @@ struct Node
     int     i, j; //grid cell coordinates
     double  F = -1, g = -1, H = -1; //f-, g- and h-values of the search node
     Node    *parent; //backpointer to the predecessor node (e.g. the node which g-value was used to set the g-velue of the current node)
-    int heapIndex;
+    size_t heapIndex;
 
     bool operator>(Node& nodeToCompare) const
     {
